pyproject.toml
[project]
name = "streaming-sts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.32.0",
    "google-genai>=1.3.0",
    "openai>=1.12.0",
    "anthropic>=0.8.0",
    "python-dotenv>=1.0.0",
    "pyaudio>=0.2.14",
    "numpy>=1.26.4",
    "google-cloud-speech>=2.25.1"
]
